Fix new order button. Fix handleDelete function

+++ src/modules/waiter/views/WaiterOrderView.vue
@@ -18,6 +18,8 @@ const id = route.params.id as string
 const order = ref<OrderResponse | null>(null)
 
 const handleDelete = async (id: string) => {
+  if (!confirm('Are you sure you want to delete this order?')) return
+
   await ordersStore.deleteOrder(id)
   if (ordersStore.isSuccess) {
     toast.success('Order deleted successfully')
@@ -64,12 +66,16 @@ onMounted(async () => {
       <h3 class="text-lg font-bold">Items:</h3>
 
       <ul class="flex flex-col gap-3">
-        <li v-for="item in order.orderItems" :key="item.id"
-          class="flex flex-col sm:flex-row gap-3 p-5 border-b border-gray-700">
+        <li
+          v-for="item in order.orderItems"
+          :key="item.id"
+          class="flex flex-col sm:flex-row gap-3 p-5 border-b border-gray-700"
+        >
           <div class="w-full sm:w-64 h-40 overflow-hidden">
             <img
               src="https://images.unsplash.com/photo-1461023058943-07fcbe16d735?q=80&w=1769&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D"
-              class="w-full h-full object-cover object-center" />
+              class="w-full h-full object-cover object-center"
+            />
           </div>
           <div>
             <p class="text-lg font-bold">{{ item.menuItem.name }}</p>
@@ -97,8 +103,10 @@ onMounted(async () => {
       <button class="w-full p-2 text-white bg-red-600 rounded" @click="() => handleDelete(id)">
         Cancel Order
       </button>
-      <RouterLink :to="{ name: 'waiter-edit-order', params: { id: order.id } }"
-        class="w-full p-2 text-white bg-gray-600 rounded text-center">
+      <RouterLink
+        :to="{ name: 'waiter-edit-order', params: { id: order.id } }"
+        class="w-full p-2 text-white bg-gray-600 rounded text-center"
+      >
         Edit Order
       </RouterLink>
     </div>
